+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - AbstractControl now uses a weak listener on the enabled state in order to prevent memory leaks.
 ### is.codion.framework.model
 - EntityEditModel.revert() added.
+- AbstractEntityEditModel.Insert, Update and Delete, refactoring and renaming.
 ### is.codion.swing.framework.ui
 - FrameworkIcons.columns() added, used in EntityTablePanel popup menu control.
 - EntityDialogs, edit panel now used as exception handler for the add and edit actions.